+++ tests/test_checker/test_noqa.py
@@ -240,7 +240,8 @@ SHOULD_BE_RAISED = types.MappingProxyType(
         'WPS473': 0,
         'WPS474': 1,
         'WPS475': 1,
-        'WPS476': 0,  # enabled only in python 3.13+
+        'WPS476': 1,
+        'WPS477': 0,  # enabled only in python 3.13+
         'WPS500': 1,
         'WPS501': 1,
         'WPS502': 0,  # disabled since 1.0.0